// back/src/interface/enum.ts
export enum ECardType {
    DANGER = "danger",
    RESOURCE = "resource",
    TREASURE = "treasure",
}

export enum RoundType {
    START = "start",
    CARD = "card",
    END_LEAVE = "end-leave",
    END_DANGER = "end-danger",
    FINISH = "end",
}

export enum EAction {
    STAY = "stay",
    LEAVE = "leave",
    NONE = "none",
}